Fix PR-requests: move 'get_job_preview_files' to *Rule models
